feat: changed int to float

+++ internal/services/import.service.go
@@ -96,7 +96,7 @@ func (s *Services) AddSomeImportedProductsToStore(tx *gorm.DB, importData *domai
 	for _, item := range importDetails {
 		if item.AcceptedCount > 0 {
 			// add imported products to store_products
-			err = tx.Exec(storeProductQuery, importData.StoreID, item.ProductID, item.ScannedCount, item.UnitPerPack*item.ScannedCount, item.SupplyPriceVat, item.RetailPriceVat, item.Vat, item.ExpireDate, item.VatSum/float64(item.ReceivedCount), item.Id, item.SeriesNumber).Error
+			err = tx.Exec(storeProductQuery, importData.StoreID, item.ProductID, item.ScannedCount, float64(item.UnitPerPack)*item.ScannedCount, item.SupplyPriceVat, item.RetailPriceVat, item.Vat, item.ExpireDate, item.VatSum/float64(item.ReceivedCount), item.Id, item.SeriesNumber).Error
 			if err != nil {
 				s.log.Error(err)
 				return err
@@ -174,7 +174,7 @@ func (s *Services) AddAllProductsToStore(tx *gorm.DB, importData *domain.Import)
 	VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
 	for _, item := range importDetails {
 		if item.ReceivedCount > 0 {
-			err = tx.Exec(storeProductQuery, importData.StoreID, item.ProductID, item.ReceivedCount, item.UnitPerPack*item.ReceivedCount, item.SupplyPriceVat, item.RetailPriceVat, item.Vat, item.ExpireDate, item.VatSum/float64(item.ReceivedCount), item.Id, item.SeriesNumber).Error
+			err = tx.Exec(storeProductQuery, importData.StoreID, item.ProductID, item.ReceivedCount, float64(item.UnitPerPack)*item.ReceivedCount, item.SupplyPriceVat, item.RetailPriceVat, item.Vat, item.ExpireDate, item.VatSum/float64(item.ReceivedCount), item.Id, item.SeriesNumber).Error
 			if err != nil {
 				s.log.Error(err)
 				return err
